chore: remove payload from api, fix permissions
