+++ package.json
@@ -49,7 +49,7 @@
   "dependencies": {
     "@aws-cdk/aws-cognito-identitypool-alpha": "latest",
     "aws-cdk-lib": "2.148.0",
-    "cdk-nag": "2.28.157"
+    "cdk-nag": "2.28.158"
   },
   "pnpm": {
     "overrides": {
